feat: use graphql and organize code
